Added Filters, faster serial transmission
I added an 128 moving average FIR-Filter and one which cuts the last 4 Bits of the data-word. I also removed the timout in the serial transmission. Now it takes 5.81 ms (the Sample Time is 5.7ms), at  an Baudrate of 25600, to push all 3 data words to the Pc. So by now the System is able read and push the data in realtime.