polished internal data transfer and hw module initialization
the data transfer between the ADC_DelSig and the main function is now realized with an global structure with an new_data flag and an 32bit data field. It could be reused when working with one of the onboard SAR ADCs.
Hardware Module Initialization moved to an dedicated function. By doing this the main function get's a lot clearer.